Point spec:nodes tests at hieradata_aws
Given that `hieradata_aws` is current and `hieradata` is legacy, it
makes sense for the tests to run against the current data hierarchy.

This does break things for the node classes that aren't being migrated
to AWS, so this provides a way of handling those exceptions without
dropping test coverage for those node classes.