Upgrade Puppet, Hiera and Facter to latest in 3.x series
Remove all of the version pinning around the Puppet, Hiera and Facter
packages.

As Puppet 3.x is now EOL, the Debian repo that serves packages for the
legacy 3.x series of Puppet versions will not recieve any further
updates. This is a simple way of getting us up to the latest 3.x.