New script, puppet3-bootstrap (installs Puppet 3)
As we are moving away from using pre-baked Ubuntu base images with specific
Puppet versions built in (as per vCloud), and towards stock upstream
Ubuntu images (as per AWS), we need a consistent way to install Puppet
early on in the provisioning process.

Currently this is handled in AWS via some scripting in the cloud-init,
but it makes more sense to put it here, as it can also be used for
Vagrant, and we drop the requirement to maintain our own Vagrant box
(previously this mirrored the vCloud image, so made sense to do both).

We also drop the dependency on the Gov.UK Apt mirror of the Puppet Labs
repo here, and go straight to upstream.