Ensure $environment_variables is available to ERB templates in govuk_jenkins
Various ERB templates in govuk_jenkins assumed the presence of an
'$environment_variables' variable. This was defined in a faraway scope, which
doesn't fly under the Puppet 4.x parser.

The fix is rather ugly, and pending a future refactor but for now simply
reference the variable from govuk_jenkins in the class which includes the
template.