Remove anchors from our documentation on dependency management
Puppet anchors existed to work around limitations in resource relationships in
earlier versions of Puppet. These limitations have been eliminated with the
introduction of the `contain` keyword. Update our documentation to advise on
the use of `contain` instead of anchors.

Also, remove the recommendation to avoid the use of resource chaining arrows.
They can lead to more readable and concise manifests in many cases.